Let admin requests pass despite a high load (BRAINSTORM-1395)
The frontend might not pass the request to a backend since the backend has been dropped from
the service list due to high load. A direct request to the backend will however let admin
requests pass. This should be especially useful for monitoring the active requests while
the load is high.